+++ zipper/protocols/graphite/graphite_group.go
@@ -74,7 +74,7 @@ func NewWithLimiter(logger *zap.Logger, config types.BackendV2, limiter *limiter
 		protocol:             config.Protocol,
 		timeout:              *config.Timeouts,
 		maxTries:             *config.MaxTries,
-		maxMetricsPerRequest: config.MaxGlobs,
+		maxMetricsPerRequest: config.MaxBatchSize,
 
 		client:  httpClient,
 		limiter: limiter,