+++ zipper/types/backend.go
@@ -11,7 +11,7 @@ type BackendsV2 struct {
 	Timeouts                  Timeouts      `mapstructure:"timeouts"`
 	KeepAliveInterval         time.Duration `mapstructure:"keepAliveInterval"`
 	MaxTries                  int           `mapstructure:"maxTries"`
-	MaxGlobs                  int           `mapstructure:"maxGlobs"`
+	MaxBatchSize              int           `mapstructure:"maxBatchSize"`
 }
 
 type BackendV2 struct {
@@ -24,7 +24,7 @@ type BackendV2 struct {
 	KeepAliveInterval   *time.Duration `mapstructure:"keepAliveInterval"`
 	MaxIdleConnsPerHost *int           `mapstructure:"maxIdleConnsPerHost"`
 	MaxTries            *int           `mapstructure:"maxTries"`
-	MaxGlobs            int            `mapstructure:"maxGlobs"`
+	MaxBatchSize        int            `mapstructure:"maxBatchSize"`
 }
 
 func (b *BackendV2) FillDefaults() {